backstage: colocate URL validation in utils.ts and extend tests
Move isValidBackstageBaseUrl into the existing utils module instead of a new utils/ folder, and add unit tests for getConfigStore and getClusterConfig.
