+++ backstage/src/utils.ts
@@ -39,3 +39,15 @@ export function getClusterConfig(cluster: string): ClusterLevelConf | null {
   }
   return conf[cluster] || null;
 }
+
+/**
+ * Returns whether `url` is a valid http(s) base URL for a Backstage instance.
+ */
+export function isValidBackstageBaseUrl(url: string): boolean {
+  try {
+    const parsed = new URL(url);
+    return parsed.protocol === 'http:' || parsed.protocol === 'https:';
+  } catch {
+    return false;
+  }
+}
